+++ src/main/java/nextstep/common/DatabaseCleanup.java
@@ -1,4 +1,4 @@
-package nextstep.subway.common;
+package nextstep.common;
 
 import com.google.common.base.CaseFormat;
 import java.util.List;
