+++ src/main/java/nextstep/subway/application/LineService.java
@@ -4,11 +4,11 @@ import java.util.List;
 import java.util.stream.Collectors;
 import nextstep.subway.NotFoundException;
 import nextstep.subway.domain.Line;
-import nextstep.subway.domain.LineRepository;
+import nextstep.subway.repository.LineRepository;
 import nextstep.subway.domain.Section;
-import nextstep.subway.domain.SectionRepository;
+import nextstep.subway.repository.SectionRepository;
 import nextstep.subway.domain.Station;
-import nextstep.subway.domain.StationRepository;
+import nextstep.subway.repository.StationRepository;
 import nextstep.subway.dto.LineRequest;
 import nextstep.subway.dto.LineResponse;
 import nextstep.subway.dto.SectionRequest;
