+++ src/main/java/nextstep/subway/application/StationService.java
@@ -3,7 +3,7 @@ package nextstep.subway.application;
 import java.util.List;
 import java.util.stream.Collectors;
 import nextstep.subway.domain.Station;
-import nextstep.subway.domain.StationRepository;
+import nextstep.subway.repository.StationRepository;
 import nextstep.subway.dto.StationRequest;
 import nextstep.subway.dto.StationResponse;
 import org.springframework.stereotype.Service;
