+++ src/test/java/nextstep/subway/acceptance/LineAcceptanceTest.java
@@ -1,14 +1,14 @@
-package nextstep.subway.station;
+package nextstep.subway.acceptance;
 
-import static nextstep.subway.station.StationAcceptanceTest.지하철역을_생성한다;
+import static nextstep.subway.acceptance.StationAcceptanceTest.지하철역을_생성한다;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import java.util.List;
+import nextstep.common.AcceptanceTest;
 import nextstep.common.RestAssuredTemplate;
 import nextstep.subway.dto.LineRequest;
-import nextstep.subway.AcceptanceTest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
